+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # sunset Changelog
 
 ## [Unreleased]
+### Added
+- Added tooltip for the random strategy in the settings page.
 
 ## [0.0.3] - 2025-07-22
 
@@ -10,6 +12,3 @@
 
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 - Add theme changing strategies
-
-[Unreleased]: https://github.com/HaMatthias/sunset/compare/v0.0.3...HEAD
-[0.0.3]: https://github.com/HaMatthias/sunset/commits/v0.0.3